testing

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@
 int main(int argc, char *argv[])
 {
 
-unsigned int mul;
+unsigned long mul;
 int i, j;
 
 	if (argc != 3)
@@ -30,8 +30,8 @@ int i, j;
 		}
 
 	}
-	mul = atoi(argv[1]) *  atoi(argv[2]);
-	printf("%d\n", mul);
+	mul = atol(argv[1]) *  atol(argv[2]);
+	printf("%lu\n", mul);
 
 return (0);
 }
